Move sliver_utils to test/widgets to fix layering
The previous commit replaced a legacy `material.dart` import in
`test/rendering/sliver_utils.dart` with `widgets.dart`. Reviewer feedback
flagged that a file under `test/rendering/` should not depend on the
widgets layer.

Since `MockSliverToBoxAdapter` extends `SingleChildRenderObjectWidget`
the widgets dependency is structural. Both consumers
(`sliver_main_axis_group_test.dart`, `sliver_cross_axis_group_test.dart`)
already live under `test/widgets/`, so this moves the utility to sit
next to them and updates the import paths.

No behavior change.
