+++ packages/flutter/test/widgets/sliver_main_axis_group_test.dart
@@ -6,9 +6,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../rendering/sliver_utils.dart';
 import 'list_tile_tester.dart';
 import 'semantics_tester.dart';
+import 'sliver_utils.dart';
 
 const double VIEWPORT_HEIGHT = 600;
 const double VIEWPORT_WIDTH = 300;
